pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homelab-infra"
version = "0.1.0"
description = "Homelab infrastructure as code"
readme = "README.md"
authors = [
    {name = "Samuel Prime", email = "sprime01@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "ansible>=9.0.0",
    "ansible-lint>=24.0.0",
    "netaddr>=1.0.0",
    "jmespath>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "molecule>=6.0.2",
    "molecule-docker>=2.1.0",
    "docker>=6.1.3",
    "requests>=2.31.0",
    "selinux>=0.3.0",
    "pytest-testinfra>=9.0.0",
    "python-loki>=1.2.3",
    "loguru>=0.7.2",
    "pytest-html>=3.2.0",
    "pytest-xdist>=3.3.1",
    "coverage>=7.2.7",
    "prometheus-client>=0.17.1",  # Added for Prometheus metrics
    "grafanalib>=0.7.0",         # For dashboard creation
]

[tool.setuptools]
packages = []

[dependency-groups]
dev = [
    "docker>=7.1.0",
    "molecule>=25.4.0",
    "molecule-docker>=2.1.0",
    "requests>=2.32.3",
    "pytest-testinfra>=9.0.0",
    "python-loki>=1.2.3",
    "pulumi>=3.94.0",
    "junit-xml>=1.9",
    "pytest-html>=3.2.0",
    "pytest-xdist>=3.3.1",
]

[tool.pytest.ini_options]
testpaths = ["tests", "ansible/roles"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
junit_family = "xunit2"
addopts = "--html=report.html --self-contained-html"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
markers = [
    "loki: tests that require Loki integration",
    "prometheus: tests that require Prometheus integration",
    "metrics: tests that generate or validate metrics"
]

[tool.molecule]
driver = "docker"
log = true
parallel = true
